requirements.txt
boto3
colorlog
schema
pyyaml
requests
deep_merge
